requirements.txt
sounddevice
scipy
numpy
faster-whisper
torch
elevenlabs
deep-translator
python-dotenv